LayoutTests/streams/readablestream-cancel-promise-rejected-async-expected.txt

ReadableStream onCancel returns a promise that will be rejected asynchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS hasRejectedSourceCancelPromise is true
PASS recordedError is errorInCancel
PASS successfullyParsed is true

TEST COMPLETE
